Move Serving scheme registration to `pkg/apis/serving/v1alpha`
