+++ pkg/apis/serving/group.go
@@ -14,11 +14,5 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package apis
-
-import "github.com/knative/eventing-sources/pkg/apis/sources/v1alpha1"
-
-func init() {
-	// Register the types with the Scheme so the components can map objects to GroupVersionKinds and back
-	AddToSchemes = append(AddToSchemes, v1alpha1.SchemeBuilder.AddToScheme)
-}
+// Package serving contains serving API versions
+package serving
