+++ pkg/apis/serving/v1alpha1/addtoscheme_serving_v1alpha1.go
@@ -14,11 +14,14 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package apis
+package v1alpha1
 
-import "github.com/knative/eventing-sources/pkg/apis/sources/v1alpha1"
+import (
+	"github.com/knative/eventing-sources/pkg/apis"
+	servingv1alpha1 "github.com/knative/serving/pkg/apis/serving/v1alpha1"
+)
 
 func init() {
 	// Register the types with the Scheme so the components can map objects to GroupVersionKinds and back
-	AddToSchemes = append(AddToSchemes, v1alpha1.SchemeBuilder.AddToScheme)
+	apis.AddToSchemes = append(apis.AddToSchemes, servingv1alpha1.SchemeBuilder.AddToScheme)
 }
